requirement.txt
requests==2.32.5
python-dotenv==1.2.1
redis==7.1.0
psycopg2-binary==2.9.11
